Split schema tests
